requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Pinecone
pinecone-client==3.0.1

# Neo4j
neo4j==5.14.1

# OpenAI (for embeddings)
openai==1.10.0

# Document Parsing
pypdf==3.17.4
python-docx==1.1.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Logging
loguru==0.7.2

# NetworkX (for graph operations)
networkx==3.2.1

# Async support
httpx==0.26.0